bugfix init_seed
