+++ easycv/apis/__init__.py
@@ -1,4 +1,3 @@
 # Copyright (c) Alibaba, Inc. and its affiliates.
 from .test import multi_gpu_test, single_cpu_test, single_gpu_test
-from .train import (build_optimizer, get_root_logger, init_random_seed,
-                    set_random_seed, train_model)
+from .train import build_optimizer, get_root_logger, train_model
