+++ easycv/datasets/loader/build_loader.py
@@ -11,6 +11,7 @@ from mmcv.runner import get_dist_info
 from torch.utils.data import DataLoader, RandomSampler
 
 from easycv.datasets.shared.odps_reader import set_dataloader_workid
+from easycv.utils.dist_utils import init_random_seed
 from easycv.utils.torchacc_util import is_torchacc_enabled
 from .collate import CollateWrapper
 from .sampler import DistributedMPSampler, DistributedSampler
@@ -58,6 +59,7 @@ def build_dataloader(dataset,
     Returns:
         DataLoader: A PyTorch dataloader.
     """
+    seed = init_random_seed(seed)
 
     if dist:
         rank, world_size = get_dist_info()
