+++ tools/prune.py
@@ -17,7 +17,7 @@ except ImportError:
         'to install pai_nni')
 
 from easycv.models import build_model
-from easycv.apis import set_random_seed, train_model, build_optimizer
+from easycv.apis import train_model, build_optimizer
 from easycv.apis.train_misc import build_yolo_optimizer
 from easycv.datasets import build_dataloader, build_dataset
 from easycv.utils.logger import get_root_logger
@@ -25,7 +25,7 @@ from easycv.file import io
 from easycv.utils.checkpoint import load_checkpoint
 from easycv.utils.config_tools import (CONFIG_TEMPLATE_ZOO,
                                        mmcv_config_fromfile, rebuild_config)
-from easycv.utils.dist_utils import get_num_gpu_per_node
+from easycv.utils.dist_utils import get_num_gpu_per_node, set_random_seed
 from easycv.toolkit.prune.prune_utils import get_prune_layer, load_pruner
